Managed to compile arm-math functions

+++ include/utils.h
@@ -12,7 +12,7 @@
 
 uint16_t int16_2_uint16(int16_t x);
 uint16_t int32_2_uint16(int32_t x);
-uint16_t mul_int16(int16_t x1,int16_t x2);
+int32_t mul_int16(int16_t x1,int16_t x2);
 
 
 #endif /* INCLUDE_UTILS_H_ */
